Added new routes for creating new uploads and viewing user's uploads. Also, finally added some error checking to the map.
