+++ app.js
@@ -13,6 +13,8 @@ var createUser = require('./routes/createUser');
 var dashboard = require('./routes/dashboard');
 var indicators = require('./routes/indicators');
 var map = require('./routes/map');
+var yourUploads = require('./routes/yourUploads');
+var createNewUpload = require('./routes/createNewUpload');
 
 var config = require('./config.js');
 
@@ -56,6 +58,8 @@ app.use('/createUser', createUser);
 app.use('/dashboard', dashboard);
 app.use('/indicators', indicators);
 app.use('/map', map);
+app.use('/yourUploads', yourUploads);
+app.use('/createNewUpload', createNewUpload);
 
 //to setup docker mysql: docker run --name episql -e MYSQL_ROOT_PASSWORD=my-secret-pw -p 3306:3306 mysql
 //this port is in use according to Docker, how can I make sure that a port is valid for me to use?
