+++ views/yourUploads.jade
@@ -1,0 +1,14 @@
+doctype html
+html
+    head
+        title Your Uploads
+        //script(async, defer, src="https://maps.googleapis.com/maps/api/js?key=AIzaSyAEKjvE48-VV37P2pGBWFphvlrx8BXGDCs&callback=initMap")
+        link(rel='stylesheet', href='/stylesheets/style.css')
+
+        script.
+            var data = !{data};
+
+    body
+
+        div Welcome to your awesome uploads page
+        div(onclick="window.location='/createNewUpload'") go to your awesome Create New Upload page
